Refactor message type handling and clean dependencies
