+++ cii-messaging-parent/cii-writer/pom.xml
@@ -39,7 +39,6 @@
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
         </dependency>
-        <!-- ADD LOMBOK DEPENDENCY -->
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
