--- cii-messaging-parent/cii-writer/src/main/java/com/cii/messaging/writer/MessageType.java
@@ -1,11 +1,0 @@
-package com.cii.messaging.writer;
-
-/**
- * Supported message types for writers.
- */
-public enum MessageType {
-    ORDER,
-    ORDER_RESPONSE,
-    DESPATCH_ADVICE,
-    INVOICE
-}
